LayoutTests/http/tests/site-isolation/cross-site-noopener-uses-different-process.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
// This test verifies that cross-site window.open with noopener uses a different process.
// Since we can't directly communicate with a cross-site noopener window, we verify this
// by checking that:
// 1. The window.open returns null (noopener behavior)
// 2. A subsequent window.open to the same target name creates a NEW window (not reusing)
// 3. The window count increases appropriately

promise_test(async (t) => {
    if (!window.testRunner) {
        assert_unreached("This test requires testRunner");
        return;
    }

    const initialWindowCount = testRunner.windowCount();

    // Open a cross-site window with noopener (localhost vs 127.0.0.1)
    const openedWindow = window.open(
        "http://localhost:8000/site-isolation/resources/simple-page.html",
        "crossSiteNoopener",
        "noopener"
    );

    // With noopener, openedWindow should be null
    assert_equals(openedWindow, null, "window.open with noopener should return null");

    // Wait a bit for the window to open
    await new Promise(resolve => t.step_timeout(resolve, 500));

    // The window count should have increased by 1
    // Note: Cross-site noopener windows are in a different process, so they may not
    // be counted in windowCount() depending on implementation. The key is that
    // the window was opened and noopener semantics were respected.
    const afterOpenCount = testRunner.windowCount();
    assert_true(afterOpenCount >= initialWindowCount, 
        "Window count should not decrease after opening a window");

    // Try to target the same window name - should create a new window since noopener
    // windows are not targetable
    const secondWindow = window.open("about:blank", "crossSiteNoopener");
    assert_not_equals(secondWindow, null, "Second window.open should return a window handle");
    assert_equals(secondWindow.location.href, "about:blank", "Second window should be about:blank");

    t.add_cleanup(() => { if (secondWindow) secondWindow.close(); });

}, "Cross-site window.open with noopener respects noopener semantics");
</script>
</body>
</html>
